Fix integer size on Windows 64bit
Patch by Chris Matradikis 2015-11-17

This version of minisat makes two assumptions:
1. long has the same size with a pointer, and
2. pointers are at least two byte aligned.

As Heinrich pointed out, the first one does not hold for 64bit Windows and the
second is not always true.

However, in practise the code works: The assumptions are made (presumably for
memory efficiency) in order to store an integer using the same memory location
with the pointer, and the least significant bit is used as a flag - if zero,
we have a pointer and if one we have an integer shifted by one bit. Even in
64bit Windows, the bits are correctly aligned since the byte order is
little-endian.

The attached patch tries to improve the situation by using size_t instead of
unsigned long in the problematic macros, which should have the correct size in
all architectures, and the check in glp_minisat1() is modified to ensure this.
Moreover, the assert that checked the pointer alignment is reinstated (again
modified to use size_t). Unfortunately, I don't have access to a 64bit Windows
system to check whether this actually helps.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>